+++ app/Http/Controllers/User/UserController.php
@@ -29,6 +29,7 @@ class UserController extends Controller
      */
     public function index(Request $request): JsonResponse
     {
+        $this->authorize('index', User::class);
         $users = $this->UserService->getUsers($request);
         return self::paginated($users, UserResource::class, 'Users retrieved successfully', 200);
 
@@ -43,6 +44,7 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request): JsonResponse
     {
+        $this->authorize('store', User::class);
         $user = $this->UserService->storeUser($request->validated());
         return self::success(new UserResource($user), 'User created successfully', 201);
     }
@@ -55,6 +57,7 @@ class UserController extends Controller
      */
     public function show(User $user): JsonResponse
     {
+        $this->authorize('show', $user);
         return self::success(new UserResource($user), 'User retrieved successfully');
     }
 
@@ -68,6 +71,7 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user): JsonResponse
     {
+        $this->authorize( 'update' , $user);
         $updatedUser = $this->UserService->updateUser($user, $request->validated());
         return self::success(new UserResource($updatedUser), 'User updated successfully');
     }
@@ -80,6 +84,7 @@ class UserController extends Controller
      */
     public function destroy(User $user): JsonResponse
     {
+        $this->authorize( 'delete' , $user);
         $user->delete();
         return self::success(null, 'User deleted successfully');
     }
@@ -91,6 +96,7 @@ class UserController extends Controller
      */
     public function showDeleted(): JsonResponse
     {
+        $this->authorize('showDeleted', User::class);
         $users = User::onlyTrashed()->get();
      return self::success(UserResource::collection($users), 'Users retrieved successfully');
 
@@ -103,6 +109,7 @@ class UserController extends Controller
      */
     public function restoreDeleted($id): JsonResponse
     {
+        $this->authorize('restoreDeleted', User::class);
         $user = User::onlyTrashed()->findOrFail($id);
         $user->restore();
         return self::success( null , 'User restored successfully');
@@ -115,10 +122,11 @@ class UserController extends Controller
      */
     public function forceDeleted($id): JsonResponse
     {
+        $this->authorize('forceDeleted', User::class);
         $user = User::onlyTrashed()->findOrFail($id)->forceDelete();
         return self::success(null, 'User force deleted successfully');
     }
-
+  
      /**
      * Calculate the average total price of all delivered orders for the user.
      *
